Do more in the base image
In general we don't need to wait for a dist-upgrade for every VM and
container.  Do that in the base image.  Also, preload the Photon
libraries and their dependencies if not any of the actual binary
packages.

Note that this change does make the test environment less true to
reality, but it seems unlikely that this will matter and the time/disk
savings are quite significant.