Add dry_run flag to setup_task
* use in vmware_carthage

* Fix a few cases where mob of none is not handled correctly

This is not an entire solution.  There are cases where a task wants to
do some thing even in dry run mode.  As an example it would be nice to
resolve the NetworkConfig so we enumerate the PortGroups involved even
if we're going to do nothing with them.

It may also be that certain levels of dampness are required.  I think
that setting config.tasks.dry_run at different levels in the injector
tree may handle that.