add_provider does not auto replace factories
There are certain classes of objects that are replaced as they are
instantiated, for example classes are replaced with instances.  The
add_provider function previously allowed such dependency providers to
be replaced even when replace was False.  This produced mysterious and
hard to debug errors.  For example, the vault.hadronindustries.com
produced by the database builder replaces the vault added into the
base injector, because the when_needed object is a factory and thus
can be replaced.

In the cases where this replacement is desired, you tend to know that you want it.
So require those callers to explicitly enable replacement.