Permit kwargs of None
Previously if a dependency kwarg was set to None, we'd call
add_provider(dependency, None).

That would end up actually performing
add_provider(InjectionKey,Dependency) because of the single-argument
form of add_provider.

Never swap the key and value if the key is an InjectionKey in
add_provider.  This will now set the dependency to be provided by an
explicit None, which will eventually get None passed into the called
function.