Split async_resolve from async_ready
There were two parts to the async_ready protocol.  The second put an
object into a ready state--running setup_tasks etc.  The first allowed
an object to be entirely replaced when it was necessary.

Split into async_ready and async_resolve.

Track ready_state so that we can try again to make an object ready if
it fails the first time.