Always use the exact argument given in injector override
If we do something like await ainjector(func, arg = foo)

func expects to receive foo as an argument, even if an injector would consider foo a factory, or even if foo is not ready.

This came up in the bootstrapping between SshKey and SshAgent, where
the key requests an agent, passing itself (a not-ready key) in
explicitly.  The key will not become ready until the agent does, but
the agent depends on the key.

For now, if someone queries the key from the subinjector, normal
resolution will happen.