Explicit VRF handling
By Linux 5.10, you must be part of a VRF in order to use it.  Simply
redirecting to the routing table does not work for TCP because the
return packets are parsed in a different context and assumed to be
part of the VRF, but the outgoing packets (at least for a local
service) are not.

So add an ssh_origin_vrf and use it explicitly.