New Plugin implementation
A plugin implementation that allows plugins to be specified either by directory or by package

Plugins consist of one or both of:

* A python module containing a function called carthage_plugin
* a file carthage_plugin.yml containing metadata

Plugins loaded from a directory must have carthage_plugin.yml at the
top of the directory.  For that reason, in addition to being searched
next to the python module specified by a plugin, carthage_plugin.yml
is also searched from the location along sys.path where top level
modules would be loaded near the loaded plugin module.  That is, if
hadron.carthage is specified as a plugin, and
hadron/carthage/__init__.py is located in /code/hadron-operations,
then carthage_plugin.yml will be searched in
/code/hadron-operations/hadron/carthage/carthage_plugin.yml and
/code/hadron-operations/carthage_plugin.yml.