Network Links
The previous NetworkConfig  approach had a number of shortcomings:

* It was not extensible beyond storing a MAC address and network for
  each interface

* It could not represent production or complicated range situations
where we want to control the switch as well as the device being
plugged into the switch.

* There was no way to pre-instantiate information given a MachineModel
  to do things like produce networkd units without also instantiating
  a machine.

Replace it with a new model.

NetworkConfigs consist of link specifications.  Link specifications
are a set of key-value pairs including at a minimum, interface, mac
(None is supported for auto assignment) and net.  The values can be
injector keys or callables .

The link specification can also specify properties of the other side of a link.

Link specifications aare resolved (calling injectors etc) into a
dictionary mapping interface names to NetworkLinks.

* Fix container to use new mechanism

* fix Vm to support new mechanism

* Fix vmware.vm.Vm to support new mechanism.

Initial work suggests that the other side link handling may want a
better interface.
Also, testing is minimal.