New methods for default keys
Sometimes the natural key when an object is used with the single
argument form of Injector.add_provider is not the InjectionKey for the
type of the object.  As an example the natural key for a machine is
InjectionKey(Machine, host = machine.name).  The natural key for a
SystemDependency is similarly dependent on the dependency in question.
Provide interfaces to permit this usage:

* Injectable.default_instance_injection_key: key to use when an
  instance is injected as a dependency provider with the single
  argument form of add_provider

* INjectable.default_class_injection_key: Same when a class is
  injected as a factory with the single argument form of add_provider.