System Dependencies: dependencies at system startup
Often it is desirable to have one machine depend on another or to
depend on some configuration task.  As an example, machines within a
site typically depend on the site router.

The Machine class provides a start_dependencies entry point for this, but we want a better mechanism to provide declarative dependencies.

SystemDependencies are an abstract interface for accomplishing this.  A system dependency is globally identified by a name and is registered in an injector context.  Machines within that context filter the injector for instances of the interface and call them, providing an AsyncInjector.

MachineDependency realizes the SystemDependency interface and starts a
specific machine, optionally waiting for it to become online.

There is now an implementation of start_dependencies in the base
Machine object.  This filters for the SystemDependencies and calls
them.  Currently there is a model-specific override interface.  If
override_dependencies is True, then only dependencies directly
registered with the model are examined.
Other behaviors are documented, but testing suggests they are not desired, so be handling of override_dependencies other than True and False is expected to change.