Rework override_dependencies
The previous implementation of override_dependencies tended to include
dependencies close to the model calling start_dependencies and exclude
dependencies further away.  That's the opposite of desired behavior in
most situations.  As an example, in setting up an enclave, it is
common to desire not to include dependencies like the router or domain
controller when setting up the router or domain controller, while
including dependencies further away like mirror servers.

Rework override_dependencies to allow an exclusion zone of
dependencies.  Within that zone default dependencies are ignored, but
dependencies defined at a broader scope than that zone are still taken
into account.