Allow configurable resolve behavior
There is no convention that is correct for all situations.  If
systemd-resolved is in use, then we probably want bind-stub.  But if
it is not, then bind-host is more correct.  Bind-host will generally
be fine unless libnss-resolved is used to get to domai n routes that
are not visible through the default resolvers.