local network config
We need to know the names of BridgeNetworks that may be attached
externally so that we pick the right locally configured bridge.  This
can be accomplished in NetworkModels by setting bridge_names.

But if the local machine is being configured by Carthage you shouldn't
have to set bridge names.  Instead, Carthage can determine the bridge
names from the BridgeLinks on the network config.

* carthage.local.process_local_network_config: examine a network
  config for the local machine (via model.network_links) and attach
  bridge names to referenced bridges.

* MachineModel: If machine_type is a subclass of LocalMachine, call
  process_local_network_config in resolve_networking.

* Add tests.