Machine.ip_address becomes model aware
Previously MachineImplementation would try to copy ip_address from the
model to the machine.  This made it hard to make ip_address on a model
a property that depended on network configuration, because at least
with carthage-runner, the machine got instantiated before
resolve_networking got called.

That has been fixed.
Even so, it is valuable to make the ip_address logic more generic.

* Machine.ip_address tries to see if the model has an ip_address
  before raising NotImplementedError.  If so, it is memoized on the
  machine instance.

* MachineImplementation in the modeling classes no longer specially handles ip_address

* vm's logic to try using qemu-guest agent to look up the ip address
  prefers an IP address from the model.