dependency_injection: Ensure placement is called  for paralle/ instantiations
IN order to run a factory for a DependencyProvider only once, the
future is captured from asynchronous instantiations and is stored as
an interim result.

Injector.get_instance did not call the placement function in such a
situation so the dependency was not properly passed into the second
and subsiquent caller looking for the same object.

Explicitly add a done_callback to call placement in such a situation.
In general get_instance tries not to deal with futures, leating that to _instantiate.
However, the interim result handling is managed by get_instance, so get_instance needs to explicitly handle this situation.

Add a test case to confirm the fix.