Do not resolve networking for volume customizations
In image creation, we sometimes create an ephemeral Container to run
tasks against a volume.  The injector context may have a
network_config.  However using the network config creates
network_links to the container that will never be used.  When tasks
enumerate the network to generate addressing or DHCP configuration,
these links are likely to be included.
That is generally an error.

* In ContainerImage and ImageVolume customization handling, explicitly
set network_config to None so no network links are created.

* In resolve_networking, correctly handle NetworkConfig of None, not just a NetworkConfig that was never set.