Don't hang on pending ready resolution
If an object is in the middle of being instantiated to ready and some
other request is made for that object in a not-ready state, return the
object that is being instantiated.  This avoids cycles hanging when
one half of the cycle is not-ready and the other is ready.  Examples
include a model requesting AnsibleInventory, which requests not-ready
access to all models.

* Pass placement into _handle_async_injectable

* Place the object (recording instantiation) once resolution has
  concluded just before we call async_become_ready

* Add a test case