filter respects order
We'd like filter to respect ordering of calls to add_provider so that
we can get BaseCustomizations to instantiate in the right order.
Similar constraints are likely to apply when enumerating plugins.

So, use a dict rather than a set to track the set of keys and return a
list.