Don't return dependency info from customization tasks
Unfortunately, if we return a last_run time from customization tasks,
we get into a loop if somehow one of the customization dependencies is
run more recently than any task in the customization.  In that case
rerunning the customization doesn't actually run any of its sub tasks
and thus doesn't update the dependency information.