Fix ssh agent cleanup
Recently, I added tracking of the ssh agent process.  However, in the
close method, the condition was if self.process rather than if
self.process is not None.  The problem is that process has an __bool__
method that ends up waiting for termination.  Unfortunately ssh-agent
will never terminate.

So, use self.process is not None as the condition