Don't add config keys
In a previous implementation of the config layer, keys needed to be
injected for all the ConfigAccessors at each level of the schema.
That's still happening, even though it is not strictly needed.

Unfortunately, it means that even if you create a ConfigLayout low in
the injector hierarchy, some of the sub-schema ConfigAccessors may be
high in the injector hierarchy, and modifications made there may
escape the intended scope.

The vmware image data store may depend on this hack as might the vmware credential handling.
That is not fixed in this commit.