Refactor vmware image handling
Git rid of VmfsDatastore and Nfsdatastore.

Instead, all datastores support uploading and local_path when set.

Refactor how we interact with config to get the vm datastore and the image datastore.

Refactor how we generate a vmdk from a raw ImageVolume.

Bug fixes to get vmware.Vm able to create a VM.