Allow overriding command to generate sshfs connection for filesystem_access
As an example use-case, we want to be able to access the filesystem of
a hosted container by sshing into the containerhost and running the
sftp server in the container filesystem namespace.

The new sshfs_process_factory method wants to be async so it can
resolve other dependencies, like making sure the containerhost is
ready for ssh.