Use -P to set empty passphrase instead of empty string to stdin
This was motivated by an issue generating keys on Windows Subsystem
for Linux (WSL).  For reasons I didn't have time to explore, on WSL,
carthage.sh.ssh-keygen invokes and blocks on ssh-askpass even when
given _in = StringIO('').

I am assuming that -P is consistent across versions of ssh-keygen.  If
so, this is cleaner than using stdin independently of debugging the
issue with WSL (for the specific case of empty passphrase).