TaskMethod: hold real reference to instance
Previously, TaskMethod held a WeakProxy reference to the instance.
This caused several problems including inability of a task method to
call super() as well as task methods not working as well from the
layout console.

It looks like this was done under the assumption that TaskMethods
would somehow get memoized into dicts in instances.  This does not
happen, so it doesn't look like a circular reference will be created.
Holding a weak reference is less efficient and problematic, so hold a
real reference instead.