Update add_setup_task to new setup_task API
This updates add_setup_task to work with the changes from
331d165bce2a49ca3a7451b25dfce8f3e65b0f16.

The TaskWrapper interface now requires a description, so we add that
as an option to add_setup_task with a boring default.

The stamp property is now generated by overriding in the TaskWrapper
subclass and cached as .stamp as as @memoproperty, or set by
__set_name__ (replacing the memoproperty).

Here we special-case the kwargs processing of 'stamp'.  If 'stamp' is
provided as a kwarg, we replace the .stamp memoproperty in the same
way as __set_name__, and set .stamp_is_explicit to True.  We use
.stamp_is_explicit to ensure we haven't created a confusing situation
where both add_setup_task and __set_name__ set the stamp name.

I'm not at all sure if the deconfliction using .stamp_is_explicit is
necessary.  I'm also not sure if there isn't a better way to set
.stamp than passing it in in entirety.