Fix Event Reparenting algorithm
If you have three objects (c child of b child of a) all attached to
the same event scope, and split the event scopes at c, and later try
to split the event scops at b, you would previously get a KeyError ind
find_prune_children.

When the event scope split at c, all the information for tracking c
and its children was correctly moved to the new event scope.  However,
c was left in the set of children for the original event scope.  When
that scope further splits, the tracking info for c was expected,
because c was listed as a child of b, but was not found because it was
in its own scope.

Fix this by modifying find_prune_children to remove the pruned child
from the parent's list of children.

This also requires keeping track of the parent in EventScope.add_child.

* Add test case to reproduce the problem.