Set up container_propagations before instantiating class.
type.__new__ is responsible for calling __init_subclass__.

In the previous code class.__container_propagations__ was set up by
ModelincContainer.__new__ after the call to super (and thus after all
the init_subclass calls).  This produced incorrect results with
add_provider_after called from init_subclass.  Effectively the
propagated item would get added to the superclass's propagations, but
ignored because it was not added to the initial injections of the
superclass.

Long term the fix is probably to move things out of hidden state in
the namespace into actual variables in the namespace.  For now, move
container_propagations into an actual variable in the namespace prior
to type.__new__ getting called.