injector_xref can take None for this injector
Provide an easier syntaxt for using injector_xref to look up a key in
the current injector.  It would be even better if we had a
one-argument form that did that.  This overlaps with the functionality
of injector_access.  Injector_access is best used for the RHS of an
assignment in a class; injector_xref is best used for calls to
add_provider.  Injector_xref supports asynchronous operation while
injector_access supports __get__.