filesystem_access defaults to ssh_online=True
I had a need for a check_completed func to run with filesystem_access.
On a podman container or other container, MachineRunning should not
get called, but on something using sshfs, MachineRunning is needed.

There's logic to make that work with  FilesystemCustomization, but duplicating that in an individual check_completed_func is undesirable.
Instead, have filesystem_access run the machine if needed

* The sshfs version of filesystem_access calls machine_running with
  ssh_online=True.  I think it the past we didn't do that because
  calling ssh_online every time even after the machine was up was
  slow.  However, now, whether ssh_online has been run is cached and
  it is only re-run if needed.

* machine_running defaults to ssh_online=True but this default can be changed in an individual Machine.
* PodmanContainer defaults not to using ssh_online