Injectable: Automatically resolve multiple inheritance
If an Injectable inherits from two classes both with dependencies,
only one will be used.  The inject decorator is smart enough to
combine the two, but Injector._instantiate assumes that the class's
dependencies are fully stated.

Add an init_subclass for Injectable to handle this situation by
calling inject() with no new dependencies for all subclasses.