transclusion at instantiation time
* Remove the complicated facility where an injector could be passed
into @transclude_overrides and the model would try to look up the
transcluding class's definition as an injector_xref value.  It was
never used and was very hard to explain.

* transcluding happens at instantiation time; this makes the
documentation of @transclude_overrides clear enough that perhaps a
user could understand it and understand when it is valuable.

* Unfortunately, because of container propagation, all propagated keys
look like they are transcluded.  Keep track of keys that are not
transcluded in InjectableModel and InjectableModelType so that
propagation does not create false transclusion.

* Update tests: injector is no longer specified.

hadron-operations will need a similar change when this makes its way
to master.