modelmethod takes a namespace proxy
Modelmethods were more different than classmethods than they needed to
be.  Create ModelingNamespaceProxy to make a ModelingNamespace look
like a class so that modelmethods can be called after class creation
and can work approximately the same either way.

Update the existing modelmethods to the new calling convention.

Currently calling outside the class body is untested.  Tests need to
be added for that.  After this makes it way to master, the hackish
add_provider_after can be removed; a corresponding change is required
in carthage_base.