Detect multiple instantiation
If add_provider is called with a dependency provider that is already
in the initial injections of the current model hierarchy, raise a
syntax error.

In almost all cases, the correct answer is to wrap in injector_access,
but provide a force option if multiple instantiation is truly desired.

Fix the podman test layout to use injector_access.

Provide a test case for this new error.