Handle configfile-not-mounted
Adapted from a patch by Klee Dienes.

If we cannot load the layout in the carthage resource agent, return
OCF_NOT_IMPLEMENTED.  For a probe, that probably results in a no-op,
which is probably what we want.  For a stop, that probably results in
a fence.  Assuming the resource was actually running, that might well
be desirable.