Transclusion should not override value
Previously aliases in the modeling layer to a key that could be
transcluded were converted into injector_access to that key.  This
meant that the alias was still available even if the key was
transcluded and would refer to the transcluded item.

Unfortunately this tended to defeat the logic in
InjectableModel.__init__ to reuse DependencyProviders.  That might
have been fine if it also defeated the logic in
Injector.replace_provider that reuses DependencyProviders so replacing
a provided object replaces all its aliases.

This led to loops:

* an object a was added using its primary key k1

* a key k2 was in a's supplementary_injection_keys, so Injector.add_provider treated it as an alias

* k2 was also an alias at the modeling layer, so InjectableModel.__init__ called injector.replace_provider replacing k2 with an injector_access for k1

* Injector.replace_provider updates the provider for k1 and k2 with the injector_access for k1.

* Instantiating k1 creates a loop.

It's still valuable to have aliases available even when a key is
transcluded.  For all the aliased keys, explicitly add an alias in
InjectableModel.__init__ **only when the key is actually trancluded**.
We do not create a loop because in this case we are not registering a
provider for the primary key since that comes from the parent
injector.  Also, use injector_xref rather than injector_access for
better async handling.