Fix PodmanPodModel networking
Commit f55aacfaa24f0 broke networking for PodmanPodModel: because
PodmanPod is not directly a ResolvableModel (and even it it were, that
key is not propagated up), its resolve_networking was never called at
layout level resolve_networking.  Because it was never called from
do_create in PodmanPod, it was never called at all.  So PodmanPod
inside a PodmanPodModel always had empty network_links.

Because network_namespace_key is set by PodmanPodModel, it is fine
that PodmanPod resolve_networking is not called until do_create time.

* Call resolve_networking in PodmanPod do_create

* Enhance tests to correctly detect whether the right IP address is assigned to a machine in a pod.