persistent_random_mac should not make promises it cannot keep
In general, we don't have the ability to set MAC addresses on
LocalMachines.  Now that we match against mac, having an inaccurate
mac in the model is problematic.  So for most cases have
persistent_random_mac return None on LocalMachines unless there is
already a mac in the store.

If a layout plans to use SystemdNetworkMixin to create devices with
non-none local_type (bridges, bonds, vlans, etc) on a LocalMachine, it
may be valuable to assign macs to these.  In that case, use
persistent_random_mac_always.
However, in most of our current layouts, such devices are more likely to be created externally and pushed to the local machine.