cross_object_dependency: support looking up on related object
Support looking up the cross_object_dependency with a getattr call as
well as having a task passed in.  It's hard to get at tasks that are
synthasized by the modeling layer except through an instance.

This will for example allow an install_mako_task for a carthage config file to be rerun when the layout is updated.