Track failed instantiations
Two new dictionaries are created: failed_instantiation_leaves which
tracks the leaf-most instantiation failures (roughly the actual
errors) and all_instantiation_failures, which tracks all failures.

In both dictionaries, keys are a tuple containing the id() of the
injector where the failure happened as well as the InjectionKey in
that injector that failed.