Use a more useful filename for compile()
As written, the script will get marked as resources/runner_console.py,
leading to very confusing output in the debugger.

I'm confident in the try: part but not sure what's best if a filename
is unavailable.  There's always '<string>' but I didn't look to see if
pdb can read from resources.

Based on your feedback I'll update the patch and the message