Convert setup.py to pyproject.toml
Based off GH-54, but we do not convert things to entrypoints, so we
preserve the ability to run out of a source tree.

An alternative would be to convert to entry points, but to leave
scripts in the source tree that adjust sys.path and then call the
entry point.
I did not go that far in this commit.