Fix up propagation of subclasses
A subclass of a ModelingContainer is a kind of template.  Imagine a
NetworkModel that includes a NetworkConfig.  When that NetworkModel is
subclassed, the NetworkConfig should be made available in the scope of
the injector of any instances of the subclass.

That worked, but up-propagation of keys outside of instances of the
template did not work because combine_mro was misused.

* Add note about combine_mro usage in utils.py

* Add test case.

* Move collecting container_propagations into
ModelingContainer.__new__ so that container_propagations is
accumulated as the class hierarchy comes together.