Ansible: do not pull name out of hosts
When running on a single host we try to pull the host-specific task
status out of the host into the task dict.  However, for task actions
like apt or pypi that include a name in the task status, this
overwrites the task name.
So, do not pull name up a level even on single hosts.