Make it possible to build the docs locally.
Created a setup_docs.sh script which can be run in side the test container to build the docs locally.
Modified the api/modeling.rst file to remove the docstring for ModelingBase from the rst file
and put it in the implementation.py file and then ensure that the api docs can autodoc that python class.