find_orphan_deployables: handle transitive dynamic dependencies
AwsVm has a dynamic dependency of AwsSubnet for each network link.
And AwsSubnet has a dynamic dependency of  AwsVirtualPrivateCloud.

Recurse into dynamic dependencies.  The loop is guaranteed to stop
assuming that all the functions it calls terminate and do not keep
generating new layers of dynamic dependencies themselves.  An object
can only contribute to a future round of the loop if when it is
processed, it has never been seen before.

* Mark dynamic dependencies readonly if they are not already ready.

* Mark orphans readonly before  calling find_deployable