Only create custom fields when trying to write them
Before this change, a VMware connection would throw an error trying to
create the Carthage fields.

After the change, it will only try to create the fields when asked to
set one on an object.

A bug is that it respects the readonly property of the object when
deciding to create a field, even though the fields are global.  I
don't have an easy fix for this.