Add NET_ADMIN and NET_RAW to support debugging
This makes debugging dramatically better.

An argument against it is that it exposes more OS surface to the
container and that it shouldn't be default.

A better solution might be something like
filter_instantiate(OciCapability), with default capabilities taken
from a configvar?  But I want to get opinions before implementing.