ssh: ssh_config for known hosts for ProxyJump
OpenSSH does not honor command line options like -oUserKnownHostsFile
for ProxyJump.  The only way to affect the options for the jump host
is to add them to ssh_config.

Carthage does not want to modify the system or user ssh_config.  So,
replace the ssh_config and try to include user and system configs as a
fallback.

* SshAgent: generate a config file.

* SshMixin: use it.

* write_ansible_config: and use it here too.

This may not pick up the ssh_config for rsync; that may need an additional change.