include_container: allow as class method
Allow include_container to be called as a class method on never
instantiated classes, so it can be called in init_subclass.

In implementing, we've decided that generally we'd like to move toward
setattr on a ModelingBase type to run the namespace filters, at least
before the class is instantiated for the first time.  That will reduce
a lot of the difference between class body and class operations and
reduce some of the magic in the modeling language.

Moving in that direction, use an ModelingNamespaceProxy in the
containment integration filter and the propagation filter.
Some of the implementation simplifications assumes that classes_to_inject will be removed as a feature in the future.