RemotePodmanHost: support runas_user != login
Support ssh_login_user != runas_user in RemotePodmanHost.  To make
this work, we need to create a directory where we can read the podman
socket from the login user, and we need to change permissions on the
socket after podman starts to make it readable.

Sockets are created in /var/lib/carthage/podman_sockets.  There is a
directory for each runas_user and in that directory, another directory
for each login user.

When memoizing container_hosts in the injector of the remote machine, cache based on host name and user, not just host name.