Fix handling of ignore_import_errors
In e81ec452791, a plugin of type module was handled by calling
importlib.import_module and passing the resulting package into
load_plugin_from_package.  Unfortunately, that ignores import_errors.

Even when there is an import error, we need to detect and try to load
the plugin enough to find its metadata so we can collect dependencies.
As a result, import_module needs to be decomposed into finding the
spec, constructing a module from the spec, and if it goes well, using
the module loader to execute the module.

That commit introdues handle_module_spec, which does everything but
finding the spec for the module.  Rather than calling import_module,
find the spec and call handle_module_spec.