find_orphan_deployables: do not leave a bunch readonly
find_orphan_deployables calls dynamic_dependencies to make sure that
it is not declaring a dynamic dependency as orphaned.  It marks the
dependencies it instantiates as readonly=DryRun, because it is
annoying if enumerating the orphans happens to create an object.
Unfortunately, if those objects do need to later get created, they are
readonly.

Modify find_orphan_deployables to remove readonly markings from any
deployables it instantiates after the orphans have been calculated.