Fix loop detection
Loops started at self.parent not at self, so the injector loop
detection handled cases where we overflowed the stack, but not cases
where we deadlocked.

Start loop detection at self.
Also, a loop where as part of asking for a ready object we ask for the same object not ready is fine.

Also implement instantiation_leaves which starts from
instantiation_roots and gives all the instantiations that are not
waiting on dependencies or that are deadlocked.