podman: commit containers on the remote host.
Podman 6.9 gets an unable to parse input stream error when it uses
podman-remote to run a commit container operation on a Podman 4.6 API
server.

Other API functions work at least for now.

We do not need local resources when running a commit, so instead just
run_command the commit operation on the container host rather than
using the podman socket.

* Add podman_nosocket operation to the container host API.

* Provide a nontrivial implementation in RemotePodmanHost

* Use this appi in PodmanImage.commit_container.