Machinemodel: explicitly propagate keys
Previously, MachineModel didn't explicitly set propagation keys.
Because it is a ModelingContainer, it must propagate, so all the keys
it provides dependencies for were chosen.  That meant that if there
was a situation like:

```python
class outer(MachineModel):
    @provides(InjectionKey('some_inner'))
    class server(MachineModel):
pass

```
Then InjectionKey('some_inner') would be propagated even though propagation was not requested.

* MachineModel: explicitly propagate MachineModel, host=cls.name and
  ResolvableModel, name=cls.name

* PodmanPodModel: propagate a ResolvableModel key and PodmanPodModel
  key.

* When mapping propagated transcluded keys, allow for the possibility
  that one of the inner keys that sholud be transcluded is dropped
  because it stops propagating.

* Add tests.